Cypress fix, rename to generic util
