+++ packages/cypress/cypress/tests/mocked/modelsAsAService/maasAuthPolicies.cy.ts
@@ -115,7 +115,7 @@ describe('MaaS Auth Policies', () => {
   it('should display the auth policies table with correct page content', () => {
     authPoliciesPage.findTitle().should('contain.text', 'Authorization policies');
     authPoliciesPage.findTable().should('exist');
-    authPoliciesPage.findRows().should('have.length', 6);
+    authPoliciesPage.findRows().should('have.length', 7);
     const premiumRow = authPoliciesPage.getRow('Premium Team Policy');
     premiumRow.findName().should('contain.text', 'Premium Team Policy');
     premiumRow.findPhase().should('contain.text', 'Ready');
@@ -137,7 +137,7 @@ describe('MaaS Auth Policies', () => {
   });
 
   it('should filter policies by keyword', () => {
-    authPoliciesPage.findRows().should('have.length', 6);
+    authPoliciesPage.findRows().should('have.length', 7);
 
     authPoliciesPage.findKeywordFilterInput().type('premium');
     authPoliciesPage.findRows().should('have.length', 1);
@@ -147,7 +147,7 @@ describe('MaaS Auth Policies', () => {
       .should('contain.text', 'Premium Team Policy');
 
     authPoliciesPage.clearAllFilters();
-    authPoliciesPage.findRows().should('have.length', 6);
+    authPoliciesPage.findRows().should('have.length', 7);
   });
 
   it('should disable the action buttons for a deleting policy in the table and view page', () => {
